app/requirements.txt
anyio==3.6.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
cssselect==1.2.0
cssutils==2.6.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
emails==0.6
exceptiongroup==1.0.4
fastapi==0.88.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
orjson==3.8.3
packaging==21.3
passlib==1.7.4
pip==22.2.2
pluggy==1.0.0
premailer==3.10.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tenacity==8.1.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wheel==0.37.1
